refactor(modules): extract platform-integration bounded context
Moves the Anti-Corruption Layer for GitLab/GitHub into src/modules/platform-integration/.

Entities (gitlab, github, diffMetadata, threadFetch):
- src/entities/<entity>/ → modules/platform-integration/entities/<entity>/

Webhook controllers + event filter:
- src/interface-adapters/controllers/webhook/
  → modules/platform-integration/interface-adapters/controllers/webhook/

Platform adapters (gitlabMergeRequest, githubPullRequest, platformAdapter):
- src/interface-adapters/adapters/
  → modules/platform-integration/interface-adapters/adapters/

Platform-specific gateways (diffMetadataFetch, threadFetch — gitlab and github):
- src/interface-adapters/gateways/<gateway>.{gitlab,github}.gateway.ts
  → modules/platform-integration/interface-adapters/gateways/

Side cleanups: normalized ~30 broken relative imports across moved
files and test files to the @/ alias.
