+++ src/modules/platform-integration/entities/diffMetadata/diffMetadata.gateway.ts
@@ -1,4 +1,4 @@
-import type { DiffMetadata } from '../reviewContext/reviewContext.js'
+import type { DiffMetadata } from '@/entities/reviewContext/reviewContext.js'
 
 export interface DiffMetadataFetchGateway {
   fetchDiffMetadata(projectPath: string, mergeRequestNumber: number): DiffMetadata
