+++ src/modules/platform-integration/interface-adapters/adapters/gitlabMergeRequest.adapter.ts
@@ -1,5 +1,5 @@
 import type { GitLabMergeRequestEvent } from '../controllers/webhook/eventFilter.js';
-import type { ReviewRequest, ReviewRequestState } from '../../entities/reviewRequest/reviewRequest.entity.js';
+import type { ReviewRequest, ReviewRequestState } from '@/entities/reviewRequest/reviewRequest.entity.js';
 
 function mapGitLabState(gitlabState: string): ReviewRequestState {
   switch (gitlabState) {
