+++ src/tests/units/interface-adapters/adapters/platformAdapter.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { PlatformAdapter } from '../../../../interface-adapters/adapters/platformAdapter.js';
+import { PlatformAdapter } from '@/modules/platform-integration/interface-adapters/adapters/platformAdapter.js';
 import { GitLabEventFactory, GitHubEventFactory } from '../../../factories/webhookEvent.factory.js';
 
 describe('PlatformAdapter', () => {
